Add generated Drift database and DAO files
Generated new Drift database and DAO files for local data storage, including app_database.g.dart and various DAO files. Updated subject_dao.dart, tables.dart, main.dart, and several model and service files to support new schema and database logic.

+++ lib/data/local/database/dao/subject_dao.dart
@@ -162,8 +162,8 @@ class SubjectDao extends DatabaseAccessor<AppDatabase> with _$SubjectDaoMixin {
   // ==================== STATISTICS ====================
 
   /// Get total credits for term
-  Future<double> getTotalCreditsForTerm(String termId) async {
+  Future<int> getTotalCreditsForTerm(String termId) async {
     final termSubjects = await getSubjectsForTerm(termId);
-    return termSubjects.fold(0.0, (sum, s) => sum + s.credits);
+    return termSubjects.fold(0, (sum, s) => sum + s.credits);
   }
 }
